--- tools/python/data-generator/requirements.txt
@@ -1,4 +1,0 @@
-pytest==8.3.4
-haversine==2.9.0
-great-circle-calculator==1.3.1
-
