+++ tools/python/sample-data-generator/requirements.txt
@@ -1,2 +1,5 @@
 boto3==1.35.51
 requests==2.32.3
+pytest==8.3.4
+haversine==2.9.0
+great-circle-calculator==1.3.1
