+++ tools/python/sample-data-generator/src/sample_data_generator/data_generator/flight_data_generator.py
@@ -1,40 +1,58 @@
-from .data_generator import DataGenerator
+import math
+import random
+import great_circle_calculator.great_circle_calculator as gcc
+from datetime import datetime, timedelta
+from haversine import haversine, Unit
+
+from .data_generator import DataGenerator, Direction
 
 class FlightDataGenerator(DataGenerator):
     def __init__(self):
+        self.airports = [
+            (33.6407, -84.4277),   # Hartsfield-Jackson Atlanta International Airport (USA)
+            (40.0799, 116.6031),   # Beijing Capital International Airport (China)
+            (25.2532, 55.3657),    # Dubai International Airport (UAE)
+            (35.5494, 139.7798),   # Tokyo Haneda Airport (Japan)
+            (1.3644, 103.9915),    # Singapore Changi Airport (Singapore)
+            (51.4700, -0.4543),    # London Heathrow Airport (UK)
+            (-33.9461, 151.1772),  # Sydney Kingsford Smith Airport (Australia)
+            (-23.4356, -46.4731),  # São Paulo/Guarulhos – Governador André Franco Montoro International Airport (Brazil)
+            (-26.1333, 28.2425),   # O.R. Tambo International Airport (South Africa)
+            (55.973927, 37.412327) # Sheremetyevo International Airport (Russia)
+        ]
+
         self.measure_templates = [
             {
                 "name": "fuel_level_gallons",
                 "type": "BIGINT",
-                "max_variation": 40,
-                "max": 8000,
+                "max_variation": 80,
+                "direction": Direction.DOWN,
+                "initial_value": 10000,
+                "max": 10000,
                 "min": 1
             },
             {
                 "name": "heading",
-                "type": "BIGINT",
-                "max_variation": 10,
-                "max": 360,
-                "min": 0
+                "type": "DOUBLE",
+                "max": 360.0,
+                "min": 0.0
             },
             {
                 "name": "air_temperature_celsius",
                 "type": "DOUBLE",
                 "max_variation": 2.0,
                 "max": 40.0,
-                "min": -90.0
+                "min": -90.0,
             },
             {
                 "name": "lat",
                 "type": "DOUBLE",
-                "max_variation": 0.5,
                 "max": 90.0,
                 "min": -90.0
             },
             {
                 "name": "lon",
                 "type": "DOUBLE",
-                "max_variation": 0.5,
                 "max": 180.0,
                 "min": -180.0
             },
@@ -42,27 +60,317 @@ class FlightDataGenerator(DataGenerator):
                 "name": "speed_knots",
                 "type": "BIGINT",
                 "max_variation": 10,
-                "max": 250,
-                "min": 200
-            }
-        ]
-        self.dimension_templates = [
-            {
-                "name": "flight_id",
-                "unique_options": ["FL123", "FL456", "FL890", "FL333", "FL100", "FL650", "FL256", "FL430", "FL211", "FL874"]
-            },
-            {
-                "name": "aircraft_type",
-                "random_options": ["Boeing 737", "Airbus A320"]
+                "max": 500,
+                "min": 400,
             }
         ]
 
-    def generate(self, start_date, end_date, reporting_frequency, num_entities, precision="MILLISECONDS", generate_unique_options_fallback=False) -> list:
+        self.dimension_templates = [
+            {
+                "name": "flight_id",
+                "unique_options": [
+                    "FL123", "FL456", "FL890", "FL333", "FL100", "FL650", "FL256", "FL430", "FL211", "FL874"
+                ]
+            },
+            {
+                "name": "aircraft_type",
+                "random_options": ["Boeing 747", "Airbus A320", "Airbus A380"]
+            }
+        ]
+
+
+    def generate(self, start_date, end_date, reporting_frequency, num_entities,
+                 precision="MILLISECONDS", generate_unique_options_fallback=False):
+        """
+        Generates time series data.
+
+        :param start_date: The start date to use when generating records. This cannot be older in hours
+            than the memory retention period in hours value for the table.
+        :param end_date: The end date to use when generating records. The maximum end date
+            Timestream for LiveAnalytics allows is 15 minutes in the future.
+        :param reporting_frequency: The frequency that records are generated by all entities, for example,
+            every 2 seconds, every 5 hours, etc.
+        :param num_entities: The number of entities that will report for each timestamp, for example,
+            the number of servers or number of stocks.
+        :param precision: The precision to use for record timestamps. Valid options are "MILLISECONDS",
+            "SECONDS", and "MICROSECONDS".
+        :param generate_unique_options_fallback: Whether to generate random strings for dimension values
+            after all values in a dimension template's "unique_options" array have been used.
+        """
         num_flight_ids = len(self.dimension_templates[0]["unique_options"])
         if num_entities > num_flight_ids and not generate_unique_options_fallback:
             raise Exception(f"num_entities ({num_entities}) was greater than the number of flight IDs ({num_flight_ids})")
-        return super().generate(start_date, end_date, reporting_frequency, num_entities, precision, generate_unique_options_fallback)
-    
+
+        # 1. Construct entities with departure & destination airports
+        self.entities = []
+        for _ in range(num_entities):
+            entity = {"latest_measures": {}}
+            for dimension_template in self.dimension_templates:
+                dimension_value_length = 10
+                if "value_length" in dimension_template:
+                    dimension_value_length = dimension_template["value_length"]
+                if "unique_options" in dimension_template:
+                    if len(dimension_template["unique_options"]) > 0:
+                        dimension_value = dimension_template["unique_options"][0]
+                        dimension_template["unique_options"].pop(0)
+                    elif generate_unique_options_fallback:
+                        dimension_value = self._generate_random_string(dimension_value_length)
+                elif "random_options" in dimension_template:
+                    dimension_value = random.choice(dimension_template["random_options"])
+
+                elif "unique_options" not in dimension_template and "random_options" not in dimension_template:
+                    dimension_value = self._generate_random_string(dimension_value_length)
+
+                entity[dimension_template["name"]] = dimension_value
+
+            # Pick departure and destination airports
+            departure_index = random.randint(0, len(self.airports) - 1)
+            destination_index = random.choice(list(set(range(len(self.airports))) - {departure_index}))
+
+            dep_lat, dep_lon = self.airports[departure_index]
+            dest_lat, dest_lon = self.airports[destination_index]
+
+            entity["current_lat"] = dep_lat
+            entity["current_lon"] = dep_lon
+            entity["target_lat"] = dest_lat
+            entity["target_lon"] = dest_lon
+
+            self.entities.append(entity)
+
+        # 2. Generate records
+        records = []
+        current_date = start_date
+        while current_date <= end_date:
+            for entity in self.entities:
+                dimensions = []
+                for k, v in entity.items():
+                    if k not in ["latest_measures", "current_lat", "current_lon",
+                                 "target_lat", "target_lon"]:
+                        dimensions.append({
+                            "Name": k,
+                            "Value": str(v),
+                            "DimensionValueType": "VARCHAR"
+                        })
+
+                measures = []
+                for measure_template in self.measure_templates:
+                    measure_name = measure_template["name"]
+                    measure_value_type = measure_template.get("type", "DOUBLE").upper()
+
+                    if measure_name in ["lat", "lon"]:
+                        measure_value = self._update_entity_position(entity, measure_name, reporting_frequency)
+                    elif measure_name == "heading":
+                        # Calculate heading based on current and target positions
+                        measure_value = self._calculate_bearing(entity["current_lat"], entity["current_lon"],
+                                                                entity["target_lat"], entity["target_lon"])
+                    else:
+                        measure_value = self._generate_measure_value(
+                            measure_template, entity, current_date == start_date
+                        )
+
+                    # Save for next iteration
+                    entity["latest_measures"][measure_name] = measure_value
+
+                    measures.append({
+                        "MeasureName": measure_name,
+                        "MeasureValueType": measure_value_type,
+                        "MeasureValue": str(measure_value)
+                    })
+
+                precision = precision.strip().upper()
+                if precision == "SECONDS":
+                    timestamp = str(int(current_date.timestamp()))
+                if precision == "MICROSECONDS":
+                    timestamp = str(int(current_date.timestamp() * 1_000_000))
+                else:
+                    # Default to millisecond precision
+                    timestamp = str(int(current_date.timestamp() * 1_000))
+
+                record = {
+                    "Dimensions": dimensions,
+                    "Time": timestamp,
+                    "Measures": measures
+                }
+                records.append(record)
+
+            current_date += reporting_frequency
+        return records
+
+    def _get_current_airport(self, lat, lon, threshold_nm=0.1):
+        """
+        Checks if the given latitude and longitude are within a threshold distance of any airport.
+        If so, returns the index of the first such airport.
+        Otherwise, returns the index of the closest airport.
+
+        :param lat: Latitude of the point in degrees.
+        :type lat: float
+        :param lon: Longitude of the point in degrees.
+        :type lon: float
+        :param threshold_nm: Distance threshold in nautical miles to consider as "at" the airport.
+        :type threshold_nm: float
+        :return: Index of the current airport if within threshold; otherwise, index of the closest airport.
+        :rtype: int
+        """
+        min_distance = float('inf')
+        closest_index = -1
+
+        for index, (airport_lat, airport_lon) in enumerate(self.airports):
+            distance = haversine((lat, lon), (airport_lat, airport_lon),  unit=Unit.NAUTICAL_MILES)
+            if distance <= threshold_nm:
+                return index
+
+            if distance < min_distance:
+                min_distance = distance
+                closest_index = index
+
+        return closest_index
+
+
+    def _update_entity_position(self, entity, measure_name, reporting_frequency, knot_speed=500):
+        """
+        Moves 'current_lat'/'current_lon' toward 'target_lat'/'target_lon' based on
+        a constant ground speed (knots) and the time delta (reporting_frequency).
+
+        If close to the target or if the distance to travel within this interval
+        exceeds the remaining distance, pick a new random airport as a destination.
+
+        :param entity: dict holding current_lat, current_lon, target_lat, target_lon, etc.
+        :param measure_name: either "lat" or "lon"
+        :param reporting_frequency: timedelta, e.g. timedelta(minutes=1)
+        :param knot_speed: flight speed in knots (nautical miles per hour)
+        :return: the updated lat or lon
+        """
+        if measure_name not in ("lat", "lon"):
+            return entity.get(f"current_{measure_name}", 0.0)
+
+        # Compute distance (in nautical miles) from current to target
+        distance_nm = haversine(
+            (entity["current_lat"], entity["current_lon"]),
+            (entity["target_lat"], entity["target_lon"]),
+            unit=Unit.NAUTICAL_MILES
+        )
+        hours_delta = reporting_frequency.total_seconds() / 3600.0
+        travel_distance_nm = knot_speed * hours_delta
+
+        # If within 0.1 nm, consider "arrived"
+        arrival_threshold_nm = 0.1
+        if distance_nm < arrival_threshold_nm or travel_distance_nm >= distance_nm:
+            entity["current_lat"] = entity["target_lat"]
+            entity["current_lon"] = entity["target_lon"]
+            # pick a new random destination
+            departure_index = self._get_current_airport(entity["current_lat"], entity["current_lon"], arrival_threshold_nm)
+            new_dest_index = random.choice(list(set(range(len(self.airports))) - {departure_index}))
+            entity["target_lat"], entity["target_lon"] = self.airports[new_dest_index]
+        else:
+            fraction = travel_distance_nm / distance_nm  # 0 < fraction < 1
+            entity["current_lon"], entity["current_lat"] = gcc.intermediate_point(
+                (entity["current_lon"], entity["current_lat"]),
+                (entity["target_lon"], entity["target_lat"]),
+                fraction
+            )
+
+        if measure_name == "lat":
+            return entity["current_lat"]
+        else:
+            return entity["current_lon"]
+
+    def _calculate_bearing(self, lat1, lon1, lat2, lon2):
+        """
+        Calculates the bearing from the first point to the second point.
+
+        :param lat1: Latitude of the starting point in degrees.
+        :type lat1: float
+        :param lon1: Longitude of the starting point in degrees.
+        :type lon1: float
+        :param lat2: Latitude of the destination point in degrees.
+        :type lat2: float
+        :param lon2: Longitude of the destination point in degrees.
+        :type lon2: float
+        :return: Bearing in degrees from north.
+        :rtype: float
+        """
+        phi1 = math.radians(lat1)
+        phi2 = math.radians(lat2)
+        delta_lambda = math.radians(lon2 - lon1)
+
+        # Normalize delta_lambda to be between -pi and +pi
+        delta_lambda = (delta_lambda + math.pi) % (2 * math.pi) - math.pi
+
+        x = math.sin(delta_lambda) * math.cos(phi2)
+        y = math.cos(phi1) * math.sin(phi2) - \
+            math.sin(phi1) * math.cos(phi2) * math.cos(delta_lambda)
+
+        theta = math.atan2(x, y)
+        bearing = (math.degrees(theta) + 360) % 360  # Normalize to 0-360
+
+        return bearing
+
+    def _generate_measure_value(self, measure_template, entity, is_first_datapoint):
+        """
+        Generates measure values for non-latitude, longitude, or heading measures based on the provided template.
+
+        :param measure_template: A dictionary containing the measure's configuration, such as name, type, 
+                                 max_variation, direction, initial_value, max, min, and random_options.
+        :type measure_template: dict
+        :param entity: The entity object containing the latest measures to base variations on.
+        :type entity: dict
+        :param is_first_datapoint: Boolean indicating whether this is the first datapoint for the entity.
+        :type is_first_datapoint: bool
+        :return: A new measure value generated based on the template and previous state.
+        :rtype: varies depending on measure type (e.g., float, int, bool, str)
+        :raises Exception: If an unknown measure_value_type is encountered in the template or prev_value is missing.
+        """
+        measure_name = measure_template["name"]
+        measure_value_type = measure_template.get("type", "DOUBLE").upper()
+        max_variation = measure_template.get("max_variation", 1.5)
+        direction = measure_template.get("direction", Direction.BIDIRECTIONAL)
+        initial_value = measure_template.get("initial_value", None)
+        max_val = measure_template.get("max", 100.0)
+        min_val = measure_template.get("min", 0.0)
+
+        if "random_options" in measure_template:
+            return random.choice(measure_template["random_options"])
+
+        # On first datapoint, pick a starting value
+        if is_first_datapoint:
+            if initial_value is not None:
+                return initial_value
+            if measure_value_type == "DOUBLE":
+                return random.uniform(min_val, max_val)
+            elif measure_value_type == "BIGINT":
+                return int(random.uniform(min_val, max_val))
+            elif measure_value_type == "BOOLEAN":
+                return random.choice([True, False])
+            elif measure_value_type == "VARCHAR":
+                return self._generate_random_string(10)
+            else:
+                raise Exception("Unknown measure_value_type")
+        else:
+            # Variation from the last known value
+            prev_value = entity["latest_measures"].get(measure_name)
+            if prev_value is None:
+                raise Exception(f"Previous value for measure '{measure_name}' not found in entity's latest_measures.")
+
+            if direction == Direction.UP:
+                variation = random.uniform(0, max_variation)
+            elif direction == Direction.DOWN:
+                variation = random.uniform(-max_variation, 0)
+            else:
+                variation = random.uniform(-max_variation, max_variation)
+
+            if measure_value_type == "DOUBLE":
+                new_value = prev_value + variation
+                return max(min_val, min(new_value, max_val))
+            elif measure_value_type == "BIGINT":
+                new_value = int(prev_value + variation)
+                return max(int(min_val), min(new_value, int(max_val)))
+            elif measure_value_type == "BOOLEAN":
+                return random.choice([True, False])
+            elif measure_value_type == "VARCHAR":
+                return self._generate_random_string(10)
+            else:
+                raise Exception("Unknown measure_value_type")
+
     def generate_dashboard(self, grafana_data_source_name, database_name, table_name) -> dict:
         return {
             "panels": [
